backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
pydantic[email]
python-dotenv
google-generativeai
motor
passlib[bcrypt]       
python-jose[cryptography] 
python-multipart